Created individual repositories for all model types
