+++ EDDiscovery/AppSettingsSecrets.config.example
@@ -2,3 +2,4 @@
   <add key="EDMaterializerUsername" value="USERNAME_HERE"/>
   <add key="EDMaterializerPassword" value="PASSWORD_HERE"/>
 </appSettings>
+
