+++ EDDiscovery/PlanetSystems/Repositories/WorldSurvey.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace EDDiscovery2.PlanetSystems.Repositories
+{
+    public class WorldSurvey : EdMaterializer
+    {
+    }
+}
